requirements.txt
flask
tensorflow==2.12.0
opencv-python
numpy
pillow
gunicorn
werkzeug
flask-cors
